requirements.txt
pygame
pyyaml
geopy
